+++ blood-donation/app/Http/Controllers/CityController.php
@@ -9,17 +9,27 @@ class CityController extends Controller
 {
     public function index()
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to access this page');
+        }
+        else{
         $cities = City::paginate(5);
         return view('cities.index', compact('cities'));
+        }
     }
     public function store(Request $request)
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to do this');
+        }
+        else{
         $request->validate([
             'city_name' => 'required',
             'region' => 'required',
         ]);
         City::create($request->all());
         return redirect()->route('cities.index')->with('success', 'City created successfully.');
+    }
 
     }
     public function edit(string $id)
@@ -32,6 +42,10 @@ class CityController extends Controller
 
     public function update(Request $request, City $city)
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to do this');
+        }
+        else{
         $request->validate([
             'city_name' => 'required',
             'region' => 'required',
@@ -39,11 +53,17 @@ class CityController extends Controller
         $city->update($request->all());
         return redirect()->route('cities.index')->with('success', 'City updated successfully.');
     }
+    }
 
     public function destroy(City $city)
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to do this');
+        }
+        else{
         $city->delete();
         return redirect()->route('cities.index');
+    }
         
     }
     public function getCities()
