+++ blood-donation/app/Http/Controllers/UserController.php
@@ -10,10 +10,14 @@ class UserController extends Controller
 {
     public function index()
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to access this page');
+        }
+        else{
         $users = User::with('roles')->get();
         $roles = Role::all();
         return view('users.index', compact('users', 'roles'));
-
+        }
     }
     public function edit(string $id)
     {
@@ -25,12 +29,17 @@ class UserController extends Controller
 
     public function update(Request $request, string $id)
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to do this');
+        }
+        else{
         $user = User::find($id);
         $role = Role::findByName($request->role);
         $user->removeRole($user->roles->first()->name);
         $user->assignRole($request->role);
         $user->update($request->all());
         return redirect()->back()->with('success', 'User updated successfully');
+        }
     }
 
     /**
@@ -38,7 +47,12 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(!auth()->user()->hasRole('admin')){
+            return redirect()->back()->with('error', 'You are not authorized to do this');
+        }
+        else{
         $user->delete();
         return redirect()->back()->with('success', 'User deleted successfully');
+        }
     }
 }
